% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ci.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculation of Confidence Interval (CI)}
\usage{
calc_ci(
  score,
  m = NULL,
  sd = NULL,
  score_type = NULL,
  rel = 0.85,
  rtm = TRUE,
  ci = 95
)
}
\arguments{
\item{score}{A numeric value representing the score for which the confidence interval is to be calculated.}

\item{m}{A numeric value representing the normative mean score. Required if `score_type` is NULL or "custom".}

\item{sd}{A numeric value representing the standard deviation of the normative mean score. Required if `score_type` is NULL or "custom".}

\item{score_type}{A character string specifying the type of the score (e.g., "z_score", "t_score", "scaled", "iq", "sten", "percentile", "custom").
If provided as a valid type (except "custom"), `m` and `sd` are ignored and fetched from `score_params`.}

\item{rel}{A numeric value between 0 and 1 representing the reliability of the measurement. Default is 0.85.}

\item{rtm}{A logical value indicating whether to adjust for regression towards the mean (TRUE) or not (FALSE). Default is TRUE.}

\item{ci}{A numeric value representing the confidence interval percentage. Default is 95.}
}
\value{
A named numeric vector containing the lower and upper confidence interval values for the provided score.
}
\description{
This function calculates and returns the lower and upper confidence interval around the provided score.
For percentiles, it converts to z-score, calculates the CI, and converts back to percentiles.
For other score types, it uses the provided or predefined mean and standard deviation.
}
\examples{
calc_ci(score = 80, m = 70, sd = 10)
calc_ci(score = 65, score_type = "t_score", rel = 0.9, rtm = FALSE, ci = 90)
calc_ci(score = -1.25, score_type = "z_score")
calc_ci(score = 75, score_type = "percentile")
calc_ci(score = 80, score_type = "custom", m = 70, sd = 8)
}
