% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_calc.R
\name{ci_calc}
\alias{ci_calc}
\title{Calculation of Confidence Interval (CI)}
\usage{
ci_calc(M, SD, total_score, rel = 0.85, rtm = TRUE, ci = 95)
}
\arguments{
\item{M}{A numeric value representing the normative mean score.}

\item{SD}{A numeric value representing the standard deviation of the normative mean score.}

\item{total_score}{A numeric value representing the total score for which the confidence interval is to be calculated.}

\item{rel}{A numeric value between 0 and 1 representing the reliability of the measurement. Default is 0.85.}

\item{rtm}{A logical value indicating whether to adjust for regression towards the mean (TRUE) or not (FALSE). Default is TRUE.}

\item{ci}{A numeric value representing the confidence interval percentage. Default is 95.}
}
\value{
A numeric vector containing the lower and upper confidence interval values for the provided total score.
}
\description{
This function calculates and returns the lower and upper confidence interval around the provided total score.
}
\examples{
ci_calc(M = 70, SD = 10, total_score = 80)
ci_calc(M = 70, SD = 10, total_score = 80, rel = 0.9, rtm = FALSE, ci = 90)
ci_calc(M = 0, SD = 1, total_score = -1.25)
}
