% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_standard_score_all_ci.R
\name{convert_standard_score_all_ci}
\alias{convert_standard_score_all_ci}
\title{Convert a Standard Score to All Possible Standard Score Types with Confidence Intervals Calculated}
\usage{
convert_standard_score_all_ci(
  score,
  from,
  m = NULL,
  sd = NULL,
  rel = 0.85,
  rtm = TRUE,
  ci = 95
)
}
\arguments{
\item{score}{A numeric value representing the score to convert.}

\item{from}{A character string specifying the type of the input score. Must be one of
"t_score", "scaled", "iq", "sten", "stanine", "z_score", "percentile", or "custom".}

\item{m}{Optional numeric value specifying the mean for a custom distribution.
Required if `from` or any `to` is "custom".}

\item{sd}{Optional numeric value specifying the standard deviation for a custom distribution.
Required if `from` or any `to` is "custom".}

\item{rel}{A numeric value between 0 and 1 representing the reliability of the measurement. Default is 0.85.}

\item{rtm}{A logical value indicating whether to adjust for regression towards the mean (TRUE) or not (FALSE). Default is TRUE.}

\item{ci}{A numeric value representing the confidence interval percentage. Default is 95.}
}
\value{
A tibble with columns for each standard score type, their converted values, and their confidence intervals.
}
\description{
Converts a given score from one standard type to all supported standard score types,
and calculates the confidence interval for each converted score.
}
\examples{
convert_standard_score_all_ci(65, from = "t_score")
convert_standard_score_all_ci(120, from = "iq", rel = 0.9, ci = 90)
convert_standard_score_all_ci(75, from = "custom", m = 80, sd = 5)

}
