% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_ci.R
\name{eval_ci}
\alias{eval_ci}
\title{Descriptive Evaluation of a Confidence Interval}
\usage{
eval_ci(ci_lower, ci_upper, score_type, lang = "en", m = NULL, sd = NULL)
}
\arguments{
\item{ci_lower}{A numeric value representing the lower bound of the confidence interval.}

\item{ci_upper}{A numeric value representing the upper bound of the confidence interval.}

\item{score_type}{A character string specifying the type of the score (e.g., "z_score", "t_score", "iq", "sten", "stanine", "scaled", "percentile", "custom").}

\item{lang}{A character string specifying the language for the description. Default is "en".}

\item{m}{Optional numeric value specifying the mean for a custom distribution. Required if `score_type` is "custom".}

\item{sd}{Optional numeric value specifying the standard deviation for a custom distribution. Required if `score_type` is "custom".}
}
\value{
A character string providing a text interpretation of the confidence interval in the selected language.
}
\description{
This function returns a text description of the given confidence interval, indicating the range of interpretation.
The scores are first converted to Z-scores for evaluation using the PsychometricCurse::eval_score function.
}
\examples{
eval_ci(40, 60, score_type = "t_score")
eval_ci(0.32, 0.95, score_type = "z_score")
eval_ci(25, 75, score_type = "percentile", lang = "cs")
eval_ci(80, 120, score_type = "custom", m = 100, sd = 10)
}
