% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_score.R
\name{eval_score}
\alias{eval_score}
\title{Descriptive Evaluation of a Provided Standard Score Value}
\usage{
eval_score(score, score_type, lang = "en", m = NULL, sd = NULL)
}
\arguments{
\item{score}{A numeric value representing the score to be evaluated.}

\item{score_type}{A character string specifying the type of the input score. Must be one of the names in the package's `score_params` list. E.g. "t_score", "scaled", "iq", "sten", "stanine", "z_score", "percentile", or "custom".}

\item{lang}{A character string specifying the language for the description. Supported languages are for example "en" (English), "cs" (Czech), and "de" (German). Default is "en".}

\item{m}{Optional numeric value specifying the mean for a custom distribution. Required if `score_type` is "custom".}

\item{sd}{Optional numeric value specifying the standard deviation for a custom distribution. Required if `score_type` is "custom".}
}
\value{
A character string providing a text interpretation of the given score in the selected language.
}
\description{
This function returns a text description of the given score, indicating whether it is above average, below average, or within the average range.
The score is first converted to a Z-score for evaluation using the PsychometricCurse::eval_score function.
}
\examples{
eval_score(65, score_type = "t_score", lang = "en")
eval_score(120, score_type = "iq", lang = "cs")
eval_score(75, score_type = "custom", lang = "de", m = 80, sd = 5)
}
