% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_zscore.R
\name{eval_zscore}
\alias{eval_zscore}
\title{Descriptive Evaluation of a Provided Z-Score Value}
\usage{
eval_zscore(z, lang = "en")
}
\arguments{
\item{z}{A numeric value representing the Z-score to be evaluated.}

\item{lang}{A character string specifying the language for the description. Currently supported languages are "en" (English), "cs" (Czech), and "de" (German). Default is "en".}
}
\value{
A character string providing a text interpretation of the given Z-score in the selected language.
}
\description{
This function returns a text description of the given Z-score, indicating whether it is above average, below average, or within the average range.
}
\examples{
eval_zscore(-1.52)
eval_zscore(1.23, lang = "en")
eval_zscore(0.1, lang = "cs")
}
