% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rci_old.R
\name{rci_old}
\alias{rci_old}
\title{Calculation of Reliable Change Index (RCI)}
\usage{
rci_old(
  total_score,
  total_score_old,
  score_type = NULL,
  m = NULL,
  sd = NULL,
  rel = 0.85,
  rel_old = NULL,
  rtm = TRUE,
  ci = 95,
  rci_method = 1,
  lang = "en",
  time = FALSE
)
}
\arguments{
\item{total_score}{A numeric value representing the current total score.}

\item{total_score_old}{A numeric value representing the previous total score.}

\item{score_type}{An optional character string specifying the type of standard score. Supported types are "z_score", "t_score", "sten", "stanine", "iq", and "scaled". If provided, M and SD are set automatically. Default is NULL, which requires M and SD to be specified.}

\item{m}{An optional numeric value representing the normative mean score. Required only if score_type is NULL.}

\item{sd}{An optional numeric value representing the standard deviation of the normative mean score. Required only if score_type is NULL.}

\item{rel}{A numeric value between 0 and 1 representing the reliability of the measurement. Default is 0.85.}

\item{rel_old}{An optional numeric value between 0 and 1 representing the reliability of the old (different) test. If provided, the function calculates RCI for two different tests.}

\item{rtm}{A logical value indicating whether to adjust for regression towards the mean (TRUE) or not (FALSE). Default is TRUE.}

\item{ci}{A numeric value representing the confidence interval percentage. Default is 95.}

\item{rci_method}{An integer value indicating the method to calculate the standard error. Default is 1. Use 1 for Lord & Novick (1968) approach, 2 for Jacobson & Truax (1991) approach.}

\item{lang}{A character string specifying the language for the description. Currently supported languages are "en" (English) and "cs" (Czech). Default is "en".}

\item{time}{A logical value indicating whether a lower score represents a better result (TRUE for time-based tests). Default is FALSE.}
}
\value{
A list containing the evaluation result (`evaluation_result`), the tested change (`tested_change`), the standard error (of prediction or difference) with confidence interval (`standard_error_with_CI`), the regression towards the mean adjustment (`regression_towards_mean`), the confidence interval percentage (`confidence_interval`), and the RCI method used (`rci_method`). Returns NA if `total_score` or `total_score_old` is NA.
}
\description{
This function calculates and returns the reliable change index, indicating whether the change in scores is statistically significant. It handles both single and two-test scenarios based on the provided parameters and supports different standard score types.
}
\examples{
# Example with custom M and SD
rci(total_score = 85, total_score_old = 80, m = 70, sd = 10, )

# Example with score_type, no need to specify M and SD
rci(total_score = 55, total_score_old = 50, score_type = "t_score")

# Example with score_type and additional parameters
rci(total_score = 12, total_score_old = 10, score_type = "scaled", rtm = FALSE, ci = 90, rci_method = 2, lang = "cs")

# Example with time parameter for time-based test
rci(total_score = 85, total_score_old = 90, score_type = "t_score", time = TRUE)
}
