% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{score_params}
\alias{score_params}
\title{Package Data Objects}
\format{
An object of class \code{list} of length 8.
}
\usage{
score_params
}
\description{
This file contains shared data objects used across the package.
}
\keyword{datasets}
