% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_eval.R
\name{z_eval}
\alias{z_eval}
\title{Text evaluation of a provided Z-Score value}
\usage{
z_eval(z, lang = "en")
}
\value{
Text interpretation of the given Z-Score, in the selected language.
}
\description{
This returns a text description of the given Z-score
Whether it is above/below average, or within the average.
}
\examples{
z_eval(-1.52)
z_eval(1.23, lang = "en")
z_eval(0.1, lang = "cs")
}
\seealso{
[stringi::stri_length()] which this function wraps.
}
