% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_eval.R
\name{zscore_eval}
\alias{zscore_eval}
\title{Descriptive Evaluation of a Provided Z-Score Value}
\usage{
zscore_eval(z, lang = "en")
}
\arguments{
\item{z}{A numeric value representing the Z-score to be evaluated.}

\item{lang}{A character string specifying the language for the description. Currently supported languages are "en" (English), "cs" (Czech), and "de" (German). Default is "en".}
}
\value{
A character string providing a text interpretation of the given Z-score in the selected language.
}
\description{
This function returns a text description of the given Z-score, indicating whether it is above average, below average, or within the average range.
}
\examples{
zscore_eval(-1.52)
zscore_eval(1.23, lang = "en")
zscore_eval(0.1, lang = "cs")
}
