% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_standard_score.R
\name{convert_standard_score}
\alias{convert_standard_score}
\title{Convert Between Different Standard Score Types}
\usage{
convert_standard_score(score, score_type, target_type, m = NULL, sd = NULL)
}
\arguments{
\item{score}{A numeric value representing the score to convert.}

\item{score_type}{A character string specifying the type of the input score. Must be one of the names in the package's `score_params` list. E.g. "t_score", "scaled", "iq", "sten", "stanine", "z_score", "percentile", or "custom".}

\item{target_type}{A character string specifying the type of the output score.  Must be one of the names in the package's `score_params` list. E.g. "t_score", "scaled", "iq", "sten", "stanine", "z_score", "percentile", or "custom".}

\item{m}{Optional numeric value specifying the mean for a custom distribution. Required if `score_type` or `target_type` is "custom".}

\item{sd}{Optional numeric value specifying the standard deviation for a custom distribution. Required if `score_type` or `target_type` is "custom".}
}
\value{
A numeric value representing the converted score.
}
\description{
Converts a score from one standard type to another. Supported score types include t_score, scaled, iq, sten, stanine, z_score, percentile, and custom distributions defined by mean (m) and standard deviation (sd).
}
\examples{
convert_standard_score(65, score_type = "t_score", target_type = "iq")
convert_standard_score(120, score_type = "iq", target_type = "stanine")
convert_standard_score(75, score_type = "custom", target_type = "z_score", m = 80, sd = 5)
}
