% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rci.R
\name{rci}
\alias{rci}
\title{Calculation of Reliable Change Index (RCI)}
\usage{
rci(
  score,
  score_old,
  score_type = NULL,
  m = NULL,
  sd = NULL,
  rel = 0.85,
  rtm = TRUE,
  ci = 95,
  rci_method = 1,
  lang = "en",
  time = FALSE
)
}
\arguments{
\item{score}{A numeric value representing the current total score.}

\item{score_old}{A numeric value representing the previous total score.}

\item{score_type}{An optional character string specifying the type of standard score. Supported types are "z_score", "t_score", "sten", "stanine", "iq", and "scaled". If provided, M and SD are set automatically. Default is NULL, which requires M and SD to be specified.}

\item{m}{An optional numeric value representing the normative mean score. Required only if score_type is NULL.}

\item{sd}{An optional numeric value representing the standard deviation of the normative mean score. Required only if score_type is NULL.}

\item{rel}{A numeric value between 0 and 1 representing the reliability of the measurement. Default is 0.85.}

\item{rtm}{A logical value indicating whether to adjust for regression towards the mean (TRUE) or not (FALSE). Default is TRUE.}

\item{ci}{A numeric value representing the confidence interval percentage. Default is 95.}

\item{rci_method}{An integer value indicating the method to calculate the standard error. Default is 1. Use 1 for Lord & Novick (1968) approach, 2 for Jacobson & Truax (1991) approach.}

\item{lang}{A character string specifying the language for the description. For example: "en" (English), "cs" (Czech), and "de" (German). Default is "en".}

\item{time}{A logical value indicating whether a lower score represents a better result (TRUE for time-based tests). Default is FALSE.}
}
\value{
A list containing the evaluation result (`evaluation_result`), the tested change in scores (`tested_change`), the standard error (of prediction or difference) with confidence interval, used to evaluate the change in scores (`cutoff`), whether the regression towards the mean was applied (`rtm`), the confidence interval percentage used for calculations (`ci`), and the RCI method used (`rci_method`). Returns NA if `score` or `score_old` is NA.
}
\description{
This function calculates and returns the reliable change index, indicating whether the change in scores is statistically significant. It handles both single and two-test scenarios based on the provided parameters and supports different standard score types.
}
\examples{
# Example with custom M and SD
rci(score = 85, score_old = 80, m = 70, sd = 10, )

# Example with score_type, no need to specify M and SD
rci(score = 55, score_old = 50, score_type = "t_score")

# Example with score_type and additional parameters
rci(score = 12, score_old = 10, score_type = "scaled", rtm = FALSE, ci = 90, rci_method = 2, lang = "cs")

# Example with time parameter for time-based test
rci(score = 85, score_old = 90, score_type = "t_score", time = TRUE)
}
